More animations and some refactoring of Gallery
