+++ app/bibliothek/page.tsx
@@ -1,3 +1,6 @@
+'use client';
+
+import { motion } from 'framer-motion';
 import Link from 'next/link';
 
 import { LibraryCard } from '@components/library/LibraryCard';
@@ -24,36 +27,45 @@ const LibraryPage = ({
 
   return (
     <PageLayout>
-      <div className="flex flex-col sm:flex-row items-center justify-center pt-6 gap-4">
-        <div className="relative">
-          <Link
-            href={{
-              pathname: '/bibliothek/galleri',
-              query: { badge: JSON.stringify(newGalleryBadgeStrings) },
-            }}
-          >
-            <LibraryCard cardTitle="Galleri" />
+      <motion.div
+        initial={{ x: -100 }}
+        animate={{ x: 0 }}
+        transition={{ type: 'spring', stiffness: 100 }}
+        className="py-15 px-5 sm:py-5"
+      >
+        <div className="flex flex-col sm:flex-row items-center justify-center pt-6 gap-4">
+          <div className="relative">
+            <Link
+              href={{
+                pathname: '/bibliothek/galleri',
+                query: { badge: JSON.stringify(newGalleryBadgeStrings) },
+              }}
+            >
+              <LibraryCard cardTitle="Galleri" />
+            </Link>
+            {newGalleryBadgeStrings.length > 0 && (
+              <NewContentBadge text="Nyt" />
+            )}
+          </div>
+          <div className="relative">
+            <Link
+              href={{
+                pathname: '/bibliothek/breve',
+                query: { badge: JSON.stringify(newLetterBadgeStrings) },
+              }}
+            >
+              <LibraryCard cardTitle="Breve"></LibraryCard>
+            </Link>
+            {newLetterBadgeStrings.length > 0 && <NewContentBadge text="Nyt" />}
+          </div>
+          <Link href="/bibliothek/vedtaegter">
+            <LibraryCard cardTitle="Vedtægter" />
           </Link>
-          {newGalleryBadgeStrings.length > 0 && <NewContentBadge text="Nyt" />}
-        </div>
-        <div className="relative">
-          <Link
-            href={{
-              pathname: '/bibliothek/breve',
-              query: { badge: JSON.stringify(newLetterBadgeStrings) },
-            }}
-          >
-            <LibraryCard cardTitle="Breve"></LibraryCard>
+          <Link href="/bibliothek/sang">
+            <LibraryCard cardTitle="Sang" />
           </Link>
-          {newLetterBadgeStrings.length > 0 && <NewContentBadge text="Nyt" />}
         </div>
-        <Link href="/bibliothek/vedtaegter">
-          <LibraryCard cardTitle="Vedtægter" />
-        </Link>
-        <Link href="/bibliothek/sang">
-          <LibraryCard cardTitle="Sang" />
-        </Link>
-      </div>
+      </motion.div>
     </PageLayout>
   );
 };
