+++ components/ui/BottomNav.tsx
@@ -22,10 +22,6 @@ export type BadgeNotification = {
   date?: Date;
 };
 
-// type FirebaseDate = {
-//   seconds: number;
-// };
-
 // TODO: fix any Mixup between docs return type and updating types
 export interface NotificationDbType {
   id?: string;
@@ -181,160 +177,3 @@ const BottomNav = () => {
 };
 
 export default BottomNav;
-
-// const badgestring1Obj: BadgeNotification = {
-//   navBar: 'bib',
-//   library: 'brev',
-//   year: 2023,
-// };
-// const badgestring1Str = makeBadgeNotifString(badgestring1Obj);
-// const badgestring1Obj2 = createBadgeFromString(badgestring1Str);
-// console.log('badgestring1Obj', badgestring1Obj);
-// console.log('badgestring1Str', badgestring1Str);
-// console.log('badgestring1Obj2', badgestring1Obj2);
-
-// const badgestring1bObj: BadgeNotification = {
-//   navBar: 'bib',
-//   library: 'gal',
-//   gallery: 'gf',
-//   year: 2023,
-// };
-// const badgestring1bStr = makeBadgeNotifString(badgestring1bObj);
-// const badgestring1bObj2 = createBadgeFromString(badgestring1bStr);
-// console.log('badgestring1bObj', badgestring1bObj);
-// console.log('badgestring1bStr', badgestring1bStr);
-// console.log('badgestring1bObj2', badgestring1bObj2);
-
-// const badgestring1cObj: BadgeNotification = {
-//   navBar: 'bib',
-//   library: 'gal',
-//   gallery: 'events',
-//   year: 2023,
-// };
-// const badgestring1cStr = makeBadgeNotifString(badgestring1cObj);
-// const badgestring1cObj2 = createBadgeFromString(badgestring1cStr);
-// console.log('badgestring1cObj', badgestring1cObj);
-// console.log('badgestring1cStr', badgestring1cStr);
-// console.log('badgestring1cObj2', badgestring1cObj2);
-
-// const badgestring1dObj: BadgeNotification = {
-//   navBar: 'bib',
-//   library: 'gal',
-//   gallery: 'tour',
-//   year: 2023,
-// };
-// const badgestring1dStr = makeBadgeNotifString(badgestring1dObj);
-// const badgestring1dObj2 = createBadgeFromString(badgestring1dStr);
-// console.log('badgestring1dObj', badgestring1dObj);
-// console.log('badgestring1dStr', badgestring1dStr);
-// console.log('badgestring1dObj2', badgestring1dObj2);
-
-// const badgestring2Obj: BadgeNotification = {
-//   navBar: 'chat',
-//   chat: 'gen',
-// };
-// const badgestring2Str = makeBadgeNotifString(badgestring2Obj);
-// const badgestring2Obj2 = createBadgeFromString(badgestring2Str);
-// console.log('badgestring2Obj', badgestring2Obj);
-// console.log('badgestring2Str', badgestring2Str);
-// console.log('badgestring2Obj2', badgestring2Obj2);
-
-// const badgestring3Obj: BadgeNotification = {
-//   navBar: 'kort',
-// };
-// const badgestring3Str = makeBadgeNotifString(badgestring3Obj);
-// const badgestring3Obj2 = createBadgeFromString(badgestring3Str);
-// console.log('badgestring3Obj', badgestring3Obj);
-// console.log('badgestring3Str', badgestring3Str);
-// console.log('badgestring3Obj2', badgestring3Obj2);
-
-// type BadgeNotification = {
-//   navBar: BadgeNotificationNavBar;
-//   library?: BadgeNotificationLibrary;
-//   gallery?: BadgeNotificationGallery;
-//   chat?: BadgeNotificationChat;
-//   year?: number;
-//   date?: Date;
-// };
-
-// const makeBadgeNotifString = (badgeNotification: BadgeNotification) => {
-//   // kort
-//   // bib-gal-tour-(1997-2023)
-//   // bib-gal-gf-(1997-2023)
-//   // bib-gal-events-(1997-2023)
-//   // bib-gal-events-(1997-2023)
-//   // bib-bre-(1997-2023)
-//   // chat-gen
-
-//   const parts: string[] = [badgeNotification.navBar];
-
-//   if (badgeNotification.library) {
-//     parts.push(badgeNotification.library);
-//   }
-
-//   if (badgeNotification.gallery) {
-//     parts.push(badgeNotification.gallery);
-//   }
-
-//   if (badgeNotification.chat) {
-//     parts.push(badgeNotification.chat);
-//   }
-
-//   if (badgeNotification.year) {
-//     parts.push(badgeNotification.year.toString());
-//   }
-
-//   const badgeString = parts.join('-');
-//   return badgeString;
-// };
-
-// const createBadgeFromString = (
-//   inputString: string
-// ): BadgeNotification | null => {
-//   const parts = inputString.split('-');
-
-//   if (parts.length < 1) {
-//     return null;
-//   }
-
-//   const badge: BadgeNotification = {
-//     navBar: parts[0] as BadgeNotificationNavBar,
-//   };
-
-//   switch (badge.navBar) {
-//     case 'kort':
-//       break;
-
-//     case 'bib':
-//       if (parts.length < 3) {
-//         // Ensure that there are at least three parts (navBar, library, gallery)
-//         return null;
-//       }
-//       badge.library = parts[1] as BadgeNotificationLibrary;
-//       switch (badge.library) {
-//         case 'gal':
-//           badge.gallery = parts[2] as BadgeNotificationGallery;
-//           badge.year = Number(parts[3]);
-//           break;
-
-//         default:
-//           badge.year = Number(parts[2]);
-//       }
-
-//       break;
-
-//     case 'chat':
-//       badge.chat = parts[1] as BadgeNotificationChat;
-//       break;
-
-//     default:
-//       return null; // Invalid navBar value
-//   }
-
-//   return badge;
-// };
-
-// type BadgeNotificationNavBar = 'kort' | 'bib' | 'chat';
-// type BadgeNotificationLibrary = 'gal' | 'brev';
-// type BadgeNotificationGallery = 'tour' | 'gf' | 'events';
-// type BadgeNotificationChat = 'gen';
